app/requirements.txt
asgiref==3.5.2
bleach==5.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-ckeditor==6.5.1
django-extra-fields==3.0.2
django-js-asset==2.0.0
django-random-filestorage==0.1.0
django-yaturbo==1.0.1
djangorestframework==3.14.0
lxml==4.9.1
Pillow==9.2.0
pytils==0.4.1
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
webencodings==0.5.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
requests==2.28.1